+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/entity/EntityManager.java
@@ -93,9 +93,9 @@ public class EntityManager {
 
   @SubscribeEvent(priority = EventPriority.HIGH)
   public static void handleEntityJoinWorldEvent(EntityJoinWorldEvent event) {
-    // Ignore client side world.
+    // Ignore client side world and if event is canceled.
     Level level = event.getWorld();
-    if (level.isClientSide) {
+    if (level.isClientSide || event.isCanceled()) {
       return;
     }
 
@@ -111,8 +111,10 @@ public class EntityManager {
     // Skip other checks if unknown entity name, multi-part or custom entity
     if (entityName == null) {
       String entityType = entity.getType().getDescriptionId();
-      if (entityType.startsWith("entity.adhooks.")) {
+      if (CoreConstants.ADHOOKS_LOADED && entityType.startsWith("entity.adhooks.")) {
         log.debug("Ignore {} entity {} in {}", CoreConstants.ADHOOKS_NAME, entity, levelName);
+      } else if (CoreConstants.MANA_AND_ARTIFICE_LOADED && entityType.startsWith("entity.mana-and-artifice.")) {
+        log.debug("Ignore {} entity {} in {}", CoreConstants.MANA_AND_ARTIFICE_NAME, entity, levelName);
       } else if (entity.isMultipartEntity() || entityType.contains("body_part")) {
         log.debug("Ignore multipart entity {} in {}.", entity, levelName);
       } else if (entity.hasCustomName()) {