+++ player/src/main/java/de/markusbordihn/adaptiveperformancetweaksplayer/player/PlayerSimulationDistance.java
@@ -30,7 +30,8 @@ import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.server.ServerLifecycleHooks;
-
+import de.markusbordihn.adaptiveperformancetweakscore.CoreConstants;
+import de.markusbordihn.adaptiveperformancetweakscore.message.WarnMessages;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLoadEvent;
 import de.markusbordihn.adaptiveperformancetweaksplayer.Constants;
 import de.markusbordihn.adaptiveperformancetweaksplayer.config.CommonConfig;
@@ -73,6 +74,11 @@ public class PlayerSimulationDistance {
         simulationDistanceMin, simulationDistanceMax, simulationDistanceDefault,
         simulationDistanceTimeBetweenUpdates);
     setSimulationDistance(simulationDistanceMin);
+
+    if (CoreConstants.DYNVIEW_LOADED) {
+      log.warn(WarnMessages.conflictingFeaturesModWarning(CoreConstants.DYNVIEW_NAME,
+          "optimizing the player simulation distance"));
+    }
   }
 
   @SubscribeEvent